Move parameter mapping functions to separate module
